Made npm start work, by re-starting with create-react-app
