+++ react-clock-old/src/styles.css
@@ -1,0 +1,158 @@
+@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700;800;900&display=swap");
+*, *::before, *::after {
+  padding: 0;
+  margin: 0;
+  border: none;
+  -webkit-box-sizing: border-box;
+          box-sizing: border-box;
+}
+
+::-moz-selection {
+  background: black;
+  color: red;
+}
+
+::selection {
+  background: black;
+  color: red;
+}
+
+#wrapper {
+  width: 100%;
+  height: 100vh;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  background: -webkit-gradient(linear, left top, right top, from(#a7e6ff), to(#6dd6ff));
+  background: linear-gradient(to right, #a7e6ff, #6dd6ff);
+}
+
+span {
+  pointer-events: none;
+}
+
+#timer-container {
+  border: 1px solid black;
+  height: 30vh;
+  border-radius: 8px;
+  width: 45%;
+  margin: auto;
+  position: relative;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+  -webkit-box-pack: space-evenly;
+      -ms-flex-pack: space-evenly;
+          justify-content: space-evenly;
+  -webkit-box-shadow: 15px 15px 30px rgba(0, 0, 0, 0.1);
+          box-shadow: 15px 15px 30px rgba(0, 0, 0, 0.1);
+}
+
+#timer-container #increment-and-decrement-wrapper {
+  width: 27%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  -ms-flex-pack: distribute;
+      justify-content: space-around;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  font-family: 'Poppins', sans-serif;
+}
+
+#timer-container #increment-and-decrement-wrapper .val-wrapper {
+  display: -ms-grid;
+  display: grid;
+  place-items: center;
+  width: 70px;
+  font-size: 35px;
+}
+
+#timer-container #increment-and-decrement-wrapper .inc-dec-wrappers {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+#timer-container #increment-and-decrement-wrapper .change-val-btn {
+  height: 30px;
+  width: 30px;
+  border: 2px solid black;
+  cursor: pointer;
+}
+
+#timer-container #increment-and-decrement-wrapper .change-val-btn i {
+  font-size: 15px;
+}
+
+#timer-container #timer-field {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  font-family: 'Poppins', sans-serif;
+  font-size: 38px;
+  -webkit-box-flex: 2;
+      -ms-flex: 2;
+          flex: 2;
+}
+
+#timer-container #timer-controls {
+  width: 27%;
+  font-size: 40px;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  -webkit-box-pack: space-evenly;
+      -ms-flex-pack: space-evenly;
+          justify-content: space-evenly;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+#timer-container #start-stop-icon-container {
+  font-size: 50px;
+}
+
+#timer-container i {
+  cursor: pointer;
+  font-size: 50px;
+}
+
+#timer-container i:active {
+  color: limegreen;
+}
+/*# sourceMappingURL=styles.css.map */
