+++ react-clock-old/src/styles.scss
